indentation fixes on most files added demo-genesis.json (from cardano-deps) and made it default for genesis file to simplify operation added double brackets to if statements for robustness added more documentation

+++ demo/template/bin/start_nodes.sh
@@ -1,4 +1,4 @@
-#/bin/bash
+#!/usr/bin/env bash
 
 # Start Nodes - Amias Channer
 #
@@ -11,9 +11,9 @@ node_min=${1:-'1'}
 node_max=${2:-$nodes_found}
 
 
-if [ $node_max -gt $nodes_found ]; then
+if [[ $node_max -gt $nodes_found ]]; then
   echo "Cannot start $node_max nodes , only $nodes_found are configured"
-  exit
+  exit 1
 fi
 
 node_count=$((1 + $node_max - $node_min)) 
@@ -22,17 +22,20 @@ echo "Starting $node_count Nodes"
 
 for counter in $(seq $node_min $node_max)	
 do
-			echo "Starting Node $counter"
-			stub='node_'$counter
-			log=$stub'.log'
-			
-			path='../nodes/'$counter'/'
-      cd $path
-			
-			command="../../bin/jormungandr --genesis-config ../../genesis.json --config ../../config.yaml --secret $stub.xprv --storage . -vvv "
-			echo $command > launch_cmd
-			echo "Running: $command"
-			screen -dmS $stub -t $stub $command 2> $stub.log 1> $stub-error.log
-			
-			cd ../../bin		
+  echo "Starting Node $counter"
+
+  stub='node_'$counter
+  log=$stub'.log'
+  path='../nodes/'$counter'/'
+  command="../../bin/jormungandr --genesis-config ../../genesis.json --config ../../config.yaml --secret $stub.xprv --storage . -vvv "
+
+  cd $path
+  if [[ -e launch_cmd ]]; then
+    echo "$stub is already running"
+  else
+    echo $command > launch_cmd
+    echo "Running: $command"
+    screen -dmS $stub -t $stub $command 2> $stub.log 1> $stub-error.log
+  fi
+  cd ../../bin
 done
